+++ FinniversKit/Sources/Components/Callout/CalloutView.swift
@@ -46,7 +46,7 @@ public final class CalloutView: UIView {
 
     private lazy var textLabel: UILabel = {
         let label = UILabel(withAutoLayout: true)
-        label.font = UIFont(name: FontType.bold.rawValue, size: 16.0)?.scaledFont(forTextStyle: .callout)
+        label.font = UIFont.font(ofSize: 16.0, weight: .bold, textStyle: .callout)
         label.textColor = .textToast
         label.numberOfLines = 0
         return label
