+++ FinniversKit/Sources/DNA/Font/FontProvider.swift
@@ -4,6 +4,13 @@
 
 import UIKit
 
+public enum FontWeight: String {
+    case light
+    case medium
+    case regular
+    case bold
+}
+
 public protocol FontProvider {
     var title1: UIFont { get }
     var title2: UIFont { get }
@@ -16,7 +23,10 @@ public protocol FontProvider {
     var captionStrong: UIFont { get }
     var detailStrong: UIFont { get }
     var detail: UIFont { get }
+    func font(ofSize size: CGFloat, weight: FontWeight) -> UIFont
 }
+
+// MARK: - Default fonts
 
 public struct DefaultFontProvider: FontProvider {
     /// FINNTypeWebStrippet-Medium with a size of 34 scaled for UIFontTextStyle.title1
@@ -151,124 +161,36 @@ public struct DefaultFontProvider: FontProvider {
         return font.scaledFont(forTextStyle: .caption1)
     }
 
+    public func font(ofSize size: CGFloat, weight: FontWeight) -> UIFont {
+        switch weight {
+        case .light:
+            return UIFont(name: FontType.light.rawValue, size: size)!
+        case .medium:
+            return UIFont(name: FontType.medium.rawValue, size: size)!
+        case .regular:
+            return UIFont(name: FontType.regular.rawValue, size: size)!
+        case .bold:
+            return UIFont(name: FontType.bold.rawValue, size: size)!
+        }
+    }
+
     private func registerCustomFonts() {
         _ = FontRegistrator.registerCustomFonts
     }
 }
 
-enum FontType: String {
+// MARK: - FINN Font types
+
+private enum FontType: String {
     case light = "FINNTypeWebStrippet-Light"
     case medium = "FINNTypeWebStrippet-Medium"
     case regular = "FINNTypeWebStrippet-Regular"
     case bold = "FINNTypeWebStrippet-Bold"
 }
 
-@objc public extension UIFont {
-    /// ## Usage:
-    /// - It should only be used one T1 and it should be the first text element that catches the users attention.
-    /// - It shall give the user an overview of which page he or she is located.
-    class var title1: UIFont {
-        Config.fontProvider.title1
-    }
-
-    /// ## Usage:
-    /// - A page can contain multiple T2 to divide content into several sections.
-    /// - There should be a lot of space between sections to create a clear distinction on the content.
-    class var title2: UIFont {
-        Config.fontProvider.title2
-    }
-
-    /// ## Usage:
-    /// - If a T2 have more sublevels, you can use T3.
-    class var title3Strong: UIFont {
-        Config.fontProvider.title3Strong
-    }
-
-    /// ## Usage:
-    /// - If a T2 have more sublevels, you can use T3.
-    class var title3: UIFont {
-        Config.fontProvider.title3
-    }
-
-    /// ## Usage:
-    /// - This have the same size as the body text, but is always bolded to differenciate them.
-    class var bodyStrong: UIFont {
-        Config.fontProvider.bodyStrong
-    }
-
-    /// ## Usage:
-    /// - This have the same size as the body text, but is always semibolded (Regular) to differenciate them.
-    class var bodyRegular: UIFont {
-        Config.fontProvider.bodyRegular
-    }
-
-    /// ## Usage:
-    /// - Regular text below titles is called body text
-    class var body: UIFont {
-        Config.fontProvider.body
-    }
-
-    /// ## Usage:
-    /// - Used for short amount of text if neither the Body or Detail is appropriate.
-    /// - This is slightly smaller than body text.
-    class var caption: UIFont {
-        Config.fontProvider.caption
-    }
-
-    /// ## Usage:
-    /// - Used for short amount of text if neither the Body or Detail is appropriate.
-    /// - Bold version of Caption
-    /// - This is slightly smaller than body text.
-    class var captionStrong: UIFont {
-        Config.fontProvider.captionStrong
-    }
-
-    /// ## Usage:
-    /// - Used for small, bold headlines.
-    class var detailStrong: UIFont {
-        Config.fontProvider.detailStrong
-    }
-
-    /// ## Usage:
-    /// - Less important information can be shown as detail text, not for long sentences.
-    /// - This is slightly smaller than body text.
-    class var detail: UIFont {
-        Config.fontProvider.detail
-    }
-
-    func scaledFont(forTextStyle textStyle: UIFont.TextStyle) -> UIFont {
-        if Config.isDynamicTypeEnabled {
-            let fontMetrics = UIFontMetrics(forTextStyle: textStyle)
-            return fontMetrics.scaledFont(for: self)
-        } else {
-            return self
-        }
-    }
-}
-
-// MARK: - Deprecated font names
-
-extension UIFont {
-    @available(*, deprecated, message: "Use bodyStrong instead.")
-    public class var title4: UIFont {
-        return bodyStrong
-    }
-
-    @available(*, deprecated, message: "Use detailStrong instead.")
-    public class var title5: UIFont {
-        return detailStrong
-    }
-
-    @available(*, deprecated, message: "Use captionStrong instead.")
-    public class var captionHeavy: UIFont {
-        return captionStrong
-    }
-}
-
 // https://medium.com/swift-and-ios-writing/a-quick-look-at-gcd-and-swift-3-732bef6e1838
 // https://stackoverflow.com/questions/37801407/whither-dispatch-once-in-swift-3/37801408
 // Registering fonts, only once instead of each time.
-
 private final class FontRegistrator {
     static let registerCustomFonts = FontRegistrator()
     init() {
@@ -278,6 +200,8 @@ private final class FontRegistrator {
         UIFont.registerFont(with: FontType.bold.rawValue)
     }
 }
+
+// MARK: - Private extensions
 
 private extension UIFont {
     static func registerFont(with filenameString: String) {
